backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.0
openai==1.55.3
pydantic==2.9.2
python-dotenv==1.0.1
